Methods

This section includes the methods implemented in BernsteinExpansions.jl.

Methods

Implicit Form

Tensorial Form

Implicit Form

Functions to compute the implicit Bernstein form of univariate and multivariate monomials.

BernsteinExpansions.univariate — Function

univariate(m::AbstractMonomialLike, l::Integer, dom::Interval)

Compute the Bernstein coefficients of a univariate monomial over an interval.

Input

m – monomial in one variable

l – degree of the Bernstein polynomial

dom – interval domain of the Bernstein expansion

Output

An l+1-dimensional vector that corresponds to the Bernstein expansion of order l of the monomial m.

Notes

For experimental purposes, different variations of the algorithm are availble in the internal function _univariate!. By dispatching on any of the following values, you can choose between:

fastmath : Uses the @fastmath. This is the fastest implementation.

fastpow : Uses fastpow from DiffEqBase.jl. This is the second fastest implementation.

base : Uses ^ from Julia. This is the slowest implementation, but it's accuracy is guaranteed to be within an <= 1 ulp for all possible input values.

Algorithm

TODO: add description (ref Smith's PhD thesis).

BernsteinExpansions.multivariate — Function

multivariate(m::AbstractMonomialLike, l::AbstractVector{Int}, dom::IntervalBox{N}) where {N}

Compute the Bernstein coefficients of a multivariate monomial.

Input

m – monomial in several variables

l – vector of degrees of the Bernstein polynomial for each variable

dom – multi-dimensional interval domain of the Bernstein expansion

Output

A vector of vectors holding the Bernstein coefficients implicitly.

Algorithm

TODO: add description (ref Smith's PhD thesis). ```

Tensorial Form

Functions to compute the tensorial Bernstein form of univariate and multivariate monomials.

BernsteinExpansions.generate_tensor_form — Function

generate_tensor_form(B::ImplicitBernsteinForm{N}; algorithm::String="loop") where {N}

Compute the Bernstein expansion in tensorial form given its implicit form.

Input

B – Bernstein expansion in implicit form

algorithm – (optional, default: "tensor_generated") the algorithm used for this calculation, valid options are:

"tensor_naive"

"tensor_generated"

"linear_kron"

"linear_kron_single"

Output

A multi-dimensional array for the "tensor" algorithms, or a one-dimensional array for the "linear" algorithms, such that

\[ A[i_1, …, i_n] = v[i_1] ⋯ v[i_n]\]

BernsteinExpansions.multivariate_tensor — Function

multivariate_tensor(k::Vector{Int64}, l::Vector{Int64},
                    low::Vector{T}, high::Vector{T};
                    algorithm::String="tensor_generated")

Compute the Bernstein coefficients of a multivariate monomial in the tensor form.

Input

k – vector of degrees for each monomial

l – vector of degrees of the Bernstein polynomial for each monomial

low – the lower bounds of the interval where the Bernstein coefficients are computed

high – the upper bounds of the interval the Bernstein coefficients are computed

algorithm – (optional, default: "tensor_generated") the algorithm used for this calculation, see valid options in generate_tensor_form

BernsteinExpansions._tensor_generated! — Function

    _tensor_generated!(A::AbstractArray{T, N}, v::VectorOfArray) where {T, N}

Compute the Bernstein expansion in tensorial form given its implicit form, using a generated function in a nested loop.

Input

A – array

v – vector of array

Output

A multi-dimensional array containing the coefficients.

Algorithm

This implementation uses Julia's Base.Cartesian to generate a set of nested loops to compute the element $A[i_1, …, i_n]$.