Methods

Methods

This section includes the methods implemented in BernsteinExpansions.jl.

Methods

Implicit Form

Tensorial Form

Implicit Form

Functions to compute the implicit Bernstein form of univariate and multivariate monomials.

BernsteinExpansions.univariate — Function.

univariate(k::Int64, l::Int64, low::N, high::N)::Vector{N} where {N<:AbstractFloat}

Compute the Bernstein coefficients of a univariate monomial.

Input

k – degree of the given monomial

l – degree of the Bernstein polynomial

low – the lower bound of the interval where the Bernstein coefficients are computed

high – the upper bound of the interval the Bernstein coefficients are computed

Output

A vector with floating point entries containing the Bernstein coefficients.

source

univariate(k::Int64, l::Int64, low::Rational, high::Rational)::Vector{Rational}

Compute exactly the Bernstein coefficients of a univariate monomial.

Input

k – degree of the given monomial

l – degree of the Bernstein polynomial

low – the lower bound of the interval where the Bernstein coefficients are computed

high – the upper bound of the interval the Bernstein coefficients are computed

Output

A vector with rational entries containing the Bernstein coefficients.

source

BernsteinExpansions.multivariate — Function.

multivariate(k::Vector{Int64}, l::Vector{Int64},
                      low::Vector{N}, high::Vector{N})::ImplicitBernsteinForm{N} where {N<:Number}

Compute the Bernstein coefficients of a multivariate monomial.

Input

k – vector of degrees for each monomial

l – vector of degrees of the Bernstein polynomial for each monomial

low – the lower bounds of the interval where the Bernstein coefficients are computed

high – the upper bounds of the interval the Bernstein coefficients are computed

Output

A Bernstein implicit form holding the Bernstein coefficients.

Examples

julia> m = multivariate([3,2],[3,2],[1.0,2],[2,4.0]);
julia> m.array
2-element Array{Array{Float64,1},1}:
 [1.0, 2.0, 4.0, 8.0]
 [4.0, 8.0, 16.0]

source

Tensorial Form

Functions to compute the tensorial Bernstein form of univariate and multivariate monomials.

BernsteinExpansions.generate_tensor_form — Function.

generate_tensor_form(B::ImplicitBernsteinForm{N}; algorithm::String="loop") where {N}

Compute the Bernstein expansion in tensorial form given its implicit form.

Input

B – Bernstein expansion in implicit form

algorithm – (optional, default: "tensor_generated") the algorithm used for this calculation, valid options are:

"tensor_naive"

"tensor_generated"

"linear_kron"

"linear_kron_single"

Output

A multi-dimensional array for the "tensor" algorithms, or a one-dimensional array for the "linear" algorithms, such that

\[ A[i_1, …, i_n] = v[i_1] ⋯ v[i_n]\]

source

BernsteinExpansions.multivariate_tensor — Function.

multivariate_tensor(k::Vector{Int64}, l::Vector{Int64},
                    low::Vector{T}, high::Vector{T};
                    algorithm::String="tensor_generated")

Compute the Bernstein coefficients of a multivariate monomial in the tensor form.

Input

k – vector of degrees for each monomial

l – vector of degrees of the Bernstein polynomial for each monomial

low – the lower bounds of the interval where the Bernstein coefficients are computed

high – the upper bounds of the interval the Bernstein coefficients are computed

algorithm – (optional, default: "tensor_generated") the algorithm used for this calculation, see valid options in generate_tensor_form

source

BernsteinExpansions._tensor_generated! — Function.

    _tensor_generated!(A::AbstractArray{T, N}, v::VectorOfArray) where {T, N}

Compute the Bernstein expansion in tensorial form given its implicit form, using a generated function in a nested loop.

Input

A – array

v – vector of array

Output

A multi-dimensional array containing the coefficients.

Algorithm

This implementation uses Julia's Base.Cartesian to generate a set of nested loops to compute the element $A[i_1, …, i_n]$.

source